싱글톤 패턴 적용 완료
